+++ Implementations/AK-SVD-Parallel/ksvd.cpp
@@ -5,14 +5,14 @@
 KSVD::KSVD(int K, int T0, int num_iter, int batch_size, unsigned int seed)
     : K_(K), T0_(T0), num_iter_(num_iter), batch_size_(batch_size), seed_(seed) {}
 
-void KSVD::fit(const Eigen::MatrixXd& Y) {
+void KSVD::fit(const Eigen::MatrixXf& Y) {
     const int M = Y.rows();
     const int N = Y.cols();
 
     // Initialise D — N×K random normal columns, then normalise each
     std::mt19937 rng(seed_);
-    std::normal_distribution<double> dist(0.0, 1.0);
-    D = Eigen::MatrixXd(N, K_);
+    std::normal_distribution<float> dist(0.0f, 1.0f);
+    D = Eigen::MatrixXf(N, K_);
     for (int k = 0; k < K_; k++) {
         for (int n = 0; n < N; n++) D(n, k) = dist(rng);
         D.col(k).normalize();
@@ -26,17 +26,17 @@ void KSVD::fit(const Eigen::MatrixXd& Y) {
         // ---- Sparse coding (parallelised inside OMP::do_omp) ----
         omp.D = D;
         omp.do_omp();
-        Eigen::MatrixXd X = omp.X;  // M × K
+        Eigen::MatrixXf X = omp.X;  // M × K
 
         // ---- Dictionary update (sequential — matches serial AK-SVD) ----
-        Eigen::MatrixXd E = Y - X * D.transpose();  // M × N
+        Eigen::MatrixXf E = Y - X * D.transpose();  // M × N
 
         std::vector<std::vector<int>> atom_signals(K_);
         for (int m = 0; m < M; m++)
             for (int k = 0; k < K_; k++)
                 if (X(m, k) != 0.0) atom_signals[k].push_back(m);
 
-        Eigen::MatrixXd E_f(M, N);
+        Eigen::MatrixXf E_f(M, N);
 
         for (int i = 0; i < K_; i++) {
             const auto& filter = atom_signals[i];
@@ -48,8 +48,8 @@ void KSVD::fit(const Eigen::MatrixXd& Y) {
 
             for (int j = 0; j < sz; j++) E_f.row(j) = E.row(filter[j]);
 
-            Eigen::VectorXd d = (E_f.topRows(sz).transpose() * (E_f.topRows(sz) * D.col(i))).normalized();
-            Eigen::VectorXd x_new = E_f.topRows(sz) * d;
+            Eigen::VectorXf d = (E_f.topRows(sz).transpose() * (E_f.topRows(sz) * D.col(i))).normalized();
+            Eigen::VectorXf x_new = E_f.topRows(sz) * d;
 
             D.col(i) = d;
             for (int j = 0; j < sz; j++)
