+++ Implementations/AK-SVD-Parallel/ksvd.hpp
@@ -8,10 +8,10 @@ public:
     KSVD(int K, int T0, int num_iter, int batch_size = 256, unsigned int seed = 42);
 
     // Learn dictionary D from data matrix Y (M signals × N dims).
-    void fit(const Eigen::MatrixXd& Y);
+    void fit(const Eigen::MatrixXf& Y);
 
-    Eigen::MatrixXd D;     // N × K  — each column is one dictionary atom
-    Eigen::VectorXd loss;  // Frobenius reconstruction error per iteration
+    Eigen::MatrixXf D;     // N × K  — each column is one dictionary atom
+    Eigen::VectorXf loss;  // Frobenius reconstruction error per iteration
 
 private:
     int K_, T0_, num_iter_, batch_size_;
